When pressing the app icon, it toggles the hidden state of the first active window. May change this to be the last selected window.
Added an event that toggles the hidden state of a window when pressing its corresponding preview body.

Made sure the app-label for the given context reappeared when no previews remained.

Created a minimisation feature for the window, and restyled the actions to be more accomodating for more actions.

FUTURE CHANGES:
1. INSTEAD OF TOGGLING ON PREVIEW BODY PRESS, I NEED TO MAKE THE WINDOW GO TO THE TOP, AND MAKE THE PREVIEWS HIDE AGAIN AFTERWARDS

2. CHANGE THE ICON PRESSING FUNCTIONALITY.