When a window is created, an initWindowBody function is called to intiailise what is meant to be inside the body of that created window based on its context.
For the email window, we add all the starting emails that should appear. However, for this commit only the preview has been added because there are some styling issues that need to be rectified first.

Additionally, some more of the css from the Bernard branch has been isolated to css_emails.css to help organise the connections between an element and their styles.