Several things have been changed:
When a focussed window is destroyed, change the focus to the first window in the group.

Made the background smoothly transition over to the next one in the new hour. This is by utilising opacity transitions.

Made the taskbar always be on top layer.

Add the degrees symbol only when the weather is there.